goods面值类型修改

+++ src/main/java/com/fnjz/front/entity/api.goods/GoodsRestDTO.java
@@ -4,6 +4,7 @@ import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 /**   
  * @Title: Entity
@@ -25,7 +26,7 @@ public class GoodsRestDTO implements java.io.Serializable {
 	/**丰丰票值*/
 	private Integer fengfengTicketValue;
 	/**面值*/
-	private Integer faceValue;
+	private BigDecimal faceValue;
 	/**列表图片（453*327）*/
 	private String listPicture;
 
@@ -85,7 +86,7 @@ public class GoodsRestDTO implements java.io.Serializable {
 	 *@return: java.lang.Integer  面值
 	 */
 	@Column(name ="FACE_VALUE",nullable=true,precision=10,scale=0)
-	public Integer getFaceValue(){
+	public BigDecimal getFaceValue(){
 		return this.faceValue;
 	}
 
@@ -93,7 +94,7 @@ public class GoodsRestDTO implements java.io.Serializable {
 	 *方法: 设置java.lang.Integer
 	 *@param: java.lang.Integer  面值
 	 */
-	public void setFaceValue(Integer faceValue){
+	public void setFaceValue(BigDecimal faceValue){
 		this.faceValue = faceValue;
 	}
